e2e: topomgr: properly clean up pods after tests
Usually, each test runs on its own namespace, and it is thus sufficient to let k8s
perform its GC to make sure all resources used during the tests are
properly cleaned up.

For topology manager things are a bit different, because resources are
gonna be statically allocated, and are gonna be accessed exclusively.

Thus, to avoid false negatives, we need
1. to run each test serially, because the tests don't take into account
   other tests when computing the resources they need (for simplicity).
   This is already implemented
2. to make sure than each test completely cleanes up after itself before
   to complete. It is not enough to make sure the GC will eventually
   clean up the unused resources, each test must start from a clean
   slate, thus the preceding tests need to make sure to do their cleanup.
   This is what the current patch implements.

Signed-off-by: Francesco Romani <fromani@redhat.com>